Fixes for x64 - this is a combination of Kobi's work and my own.

+++ src/clientlist.c
@@ -68,7 +68,7 @@ void remove_from_list(DLink **list, void *ptr, DLink *link)
       }
    }
 
-   sendto_realops("remove_from_list(%x, %x) failed!!", (int) list, (int) ptr);
+   sendto_realops("remove_from_list(%x, %x) failed!!", (u_long) list, (u_long) ptr);
 }
 
 u_long
